+++ ilastik/experimental/parser/types/applets.py
@@ -1,7 +1,7 @@
 ###############################################################################
 #   ilastik: interactive learning and segmentation toolkit
 #
-#       Copyright (C) 2011-2024, the ilastik developers
+#       Copyright (C) 2011-2026, the ilastik developers
 #                                <team@ilastik.org>
 #
 # This program is free software; you can redistribute it and/or
@@ -21,8 +21,7 @@
 # pyright: strict
 from codecs import decode
 from collections import OrderedDict
-from pathlib import Path
-from typing import Annotated, Dict, List, Literal, Optional, Tuple, Type
+from typing import Annotated, Dict, List, Literal, Optional, Tuple
 
 import annotated_types
 import numpy
@@ -36,9 +35,9 @@ from ilastik.applets.base.appletSerializer.legacyClassifiers import (
 from ilastik.applets.base.appletSerializer.serializerUtils import deserialize_string_from_h5
 from ilastik.experimental.parser._h5helpers import (
     deserialize_arraylike_from_h5,
-    deserialize_axistags_from_h5,
     deserialize_str_list_from_h5,
 )
+from ilastik.experimental.parser.types.base import DatasetInfo
 from lazyflow.classifiers import LazyflowVectorwiseClassifierABC, LazyflowVectorwiseClassifierFactoryABC
 
 NDShape = Annotated[Tuple[int, ...], annotated_types.Len(2, 6)]
@@ -75,20 +74,6 @@ class ProjectBase(BaseModel):
         BeforeValidator(decode),
         BeforeValidator(deserialize_arraylike_from_h5),
     ] = Field(alias="workflowName")
-
-
-class VigraAxisTags(BaseModel):
-    key: str
-    typeFlags: int
-    resolution: int
-    description: str
-
-
-class DatasetInfo(BaseModel):
-    axistags: Annotated[List[VigraAxisTags], BeforeValidator(deserialize_axistags_from_h5)]
-    shape: Annotated[
-        NDShape, BeforeValidator(lambda x: tuple(x.tolist())), BeforeValidator(deserialize_arraylike_from_h5)
-    ]
 
 
 class InputData(BaseModel):
